rabbitmq: Configure epmd.socket to listen on admin IP address too
New erlang packages on openSUSE/SLES ship with a system-wide epmd.socket
unit, and when this exists, rabbitmq will just use this by default
instead of starting its own epmd.

However by default, epmd will only listen on localhost so we need to
configure it to listen on the admin IP address too.

The code is compatible with previous packages, since it will only do
this if the rabbitmq-server.service unit depends on the epmd.service
unit; when it's not the case, we know rabbitmq-server will start epmd on
its own.