rabbitmq: Use enhanced features in clustered OCF resource agent
The resource agent has some featureto require less network usage / be
more performant (for monitoring health and listing queues), but this
requires a version of rabbitmq that is recent (newer than 3.6.4, at
least).

We will have this now, but still use an attribute so we can backport the
patch safely without breaking anything on old distros (it may be needed).