rabbitmq: Do not mirror queues to all nodes
Mirroring queues is known to have an impact on performance [1] [2]. We
were looking at mirroring queues to all nodes, but what we really need
is to mirror to a quorum of nodes. This should improve things a bit as a
first step. We may choose to not mirror some queues as a later step.

[1] "To How Many Nodes to Mirror?" in https://www.rabbitmq.com/ha.html
[2] http://fuel-ccp.readthedocs.io/en/latest/design/ref_arch_1000_nodes.html#replication