neutron: skip neutron-ha-tool when using l3_ha
neutron-ha-tool helps moving routers from one node to other active
nodes. This is triggered when the L3 agent on that particular network
node is down. This behavior is not compatible with L3 HA.

L3 HA implementation uses one master router and several others in
standby. If a router/agent is down, it will be discovered through VRRP
protocol [1] and another one will take the master role.

Both L3 HA and neuton-ha-tool are incompatible because they both has
their own way of balancing routers on fallen nodes.

This implementation lets the user configure L3 HA, which will disable
neutron-ha-tool. If L3 HA is deactivated, neutron-ha-tool comes back.

Notice that switching between neutron-ha-tool and L3 HA is possible, but
the routers created are not updated to reflect this change. If they were
created as HA routers they will remain that way. The contrary is also
true. It is up to the user to manually change those routers or delete
them and create new ones.

[1] https://wiki.openstack.org/wiki/Neutron/L3_High_Availability_VRRP