rabbitmq: Start epmd.socket, don't just enable it
Otherwise, well, it's not going to be really used until a reboot.

This matters for HA, but also because in non-HA mode because otherwise,
our call to rabbitmq-plugins (before starting rabbitmq-server) will
result in epmd to be spawned without systemd knowing it, and this will
create failures for the rabbitmq-server.service unit since epmd.service
won't be able to start (the port being already used).