rabbitmq: Do not use HiPE on virtualized platforms
The slow startup it results in can be significantly slower on
virtualized platforms, and especially in HA, this can result in things
not starting on time.

Also, HiPE is there to improve performance, and the first step to
improve performance may be to move out of a virtualized server.

(cherry picked from commit baa35b79985068bae4f93e80ecbbbec21b06daa2)