rabbitmq: add some performance configs
Adds some performance configs to tweak rabbitmq:

- backlog: number of connections that can be initialized at the
same time. Useful for a high number of connections or for a temporal
high numbers of connections, like in a server restart. Default is 128,
recommended is 4096.

- sndbuf: increase buffer for TCP connections of consumers, on linux
this value is set automatically between 80KB and 120 KB.
The higher this value is, the higher the throughput and the RAM consumed.
Recommended values for this config depend on the number of connections
and the amount of RAM available.

- recbuf: same as sndbuf but for publishers and protocol operations.

- collect_statistics_interval: the interval in milliseconds to collect
statistics, mainly by the management plugin. Increasing it will reduce
the load on the node that has the stats collector DB. Default is 5000,
recommended is 30000 or even 60000, depending on how recent you need
the stats.

(cherry picked from commit 25f8f347f385594f4d8742bc898f23473304001c)