keystone: Update endpoint before chef run
Without this change we run into scenarios where if the chef run fails for any reason
when we modify the endpoints in keystone the data is not updated/saved as apply_role_post_chef_call
is only called after the chef run. To avoid this case we should save the endpoint data in
apply_role_pre_chef_call as well.