cinder: Set os_privileged_* values
When the privileged user is not set, cinder does not provide
correct credentials to nova client and migration of attached
volumes fails.

See https://bugs.launchpad.net/cinder/+bug/1614344 for upstream bug.
The relevant code was rewritten for Pike so providing those options
won't be necessary in newer releases. But as the work was not
backported, we have to provide os_privileged values in Newton codebase
if we want to make the migration work (bsc#1079763).