magnum: Remove SLES image upload (bsc#1082278)
This removes the upload of the SLES based magnum image to glance since we no longer
support this image in Cloud 8. Only the images that are supported upstream can
be used with magnum going forward. This also removes the specific flavor we
created for the mangum image since it does not make sense to leave that in
there.