neutron: [cisco-aci]: Add support for vpc_pairs.
Cisco ACI supports connecting to the fabric either through
direct port mappings that in which case, we need to specify
the exact ports under [apic-switches] configuration. The pattern
usually looks like "1/34" for interface Eth1 and Port 34. Associating
these ports creates new policy profiles and groups for the integration.
This is not feasible if the ACI is already running another production
cloud and we need to integrate openstack as a supplement (which is
the case with most ACI customers).
However, ACI also allows for associating pre-existing profiles using
VPC_Pairs. This can be configured by specifying the apic_vpc_pairs in
ml2_conf_cisco_apic.ini.
This commit provides an option to test with either modes (VPC Pairs
or direct connection) as the use-case may be.