rabbitmq: check for rabbit readiness
As the other checks are not enough, as pacemaker keeps restarting
rabbitmq, we need a more robust way of checking that rabbit has entered
an stable situation.

So check that rabbit is up 5 times in a row with a delay of 2 seconds
between checks to make sure pacemaker has left it alone.

Also, only trigger that check for rabbit if the pacemaker_transaction is
updated, otherwise there is no need to do so