nova: disable progress timeout for live migration
The progress timeout is unreliable and was already disabled
by default in pike, however it is still enabled in Newton, hence
this is going to be backported. Also be a little bit more generous
to slow compute nodes and allow migration to as slow as 1MByte/s
before aborting it. this tremendeously helps live migration to
succeed on very busy compute nodes (where migrating away busy
workload is the best thing to do)

Also remove an outdated setting that is defaulting to 30 already.

(cherry picked from commit 4e1feb7cea89092f17e838f8df3b7993ed6bc539)