nova: avoid scheduling conflicts on HA CP
When all schedulers are picking the "best" host, in a 3 node
HA case they will all pick the same host, which is suboptimal.
Try to scale out a bit and increase retries in case of conflicts.

Even worse, a node that had still capacity when the scheduling
started might be out of memory at the time the VM is launched,
causing failed starts of VMs.

(cherry picked from commit 89071b47aef48f90146713867fca1047b5766fe9)