mysql: Create clustercheck config on nodes early enough (bsc#1059530)
The config files were created very late, just before we create and start
the galera-python-clustercheck pacemaker resource. As a result, only the
founder nodes had the config files when the resource is started, which
resulted in the other nodes not having galera-python-clustercheck
listening on the right port.

Part of https://bugzilla.suse.com/show_bug.cgi?id=1059530