mysql: Ensure that the galera resource has a monitor op (bsc#1059530)
The monitor op was dropped by accident with
2c5d20d15f8e96ad206eddb04d726433b0f2615e: when we moved the definition
of the some of the ops in the proposal, what happened is that the all
hash defining the ops in default attributes has been overridden by the
hash from the proposal, hence the monitor op getting lost.

Now we merge the two to get the desired results. This makes sure that if
mariadb is stopped, pacemaker will restart it.

https://bugzilla.suse.com/show_bug.cgi?id=1059530