rabbitmq: Do not check if service is running when it is not supposed to run.
pre-upgrade attribute is set to true to indicate the case that the upgrade of the node has
not been finished yet. In such case, services cannot start on the node
(there's a location constraint that prevents that).
See disable_pre_upgrade_attribute_for method in models/api/node.rb for more info