galera: change listening address/port with haproxy
During our work to make haproxy active in all controller nodes, we found
out Galera is not listening to all interfaces (0.0.0.0) but the VIP
address. This scenario works when there's only one haproxy checking for
Galera but fails on multiple haproxy scenario.

By changing the address and port where Galera is listening we aim to
avoid the conflict for service listening.