glance: registry to use same listening addr as API
In our work to make haproxy active on all controllers we found some
services are listening on the VIP address rather than on all interfaces.
When it listens on the VIP adress and haproxy is configured in several
nodes then there's a problem binding the socket to that VIP address.

This change makes glance-registry to use a listen-to-all interfaces
(0.0.0.0) by imitating the same mechanism that works perfectly for
glance-api.