nova: metadata to listen to all interfaces
In our work to make haproxy active on all controllers we found some
services are listening on the VIP address rather than on all interfaces.
When it listens on the VIP adress and haproxy is configured in several
nodes then there's a problem binding the socket to that VIP address.

This change makes nova-metadata to use a listen-to-all interfaces
(0.0.0.0) as it is configured in all other nova services.