mysql: swap haproxy and cluster-check creation
While working in the haproxy active/active feature, we found out that
galera clustercheck script was failing on non founder nodes. We believe
the cause is clustercheck trying to get the status of the service while
haproxy has not still been activated.

By swapping clustercheck and haproxy creation we aim to give time for
clustercheck to find the service active and thus avoid errors.