keystone: avoid race condition during admin password change (bsc#1091829)
Issuing a new keystone token immediately after updating the
admin user password may sometimes return an invalid token.

In the context of crowbar, this issue can be triggered when calling
the keystone_register 'wakeup' action immediately after the admin
password has been updated. When triggered, it results in timeout errors
on non-founder nodes, while the founder node is stuck doing retry
iterations with an expired token.

As a workaround for bsc#1091829, the 'wakeup' action is updated with
an optional 'reissue_token_on_error' argument, which, when set, will
re-issue a token *before* checking the keystone API again, instead of
reusing the same token for subsequent attempts.