rabbitmq: install and configure yarb to launch automatically
When restarting a rabbit cluster usually all queues are created
with the first node to come back as the master which leads to
a severe queue unbalance between the nodes and extra load on
just one of them.

This commit makes it so we can install and configure yarb to
balance the rabbit queues automagically by using ClusterMon
to catch cluster actions and react to any rabbitmq
coming back from a restart or promotion to launch yarb.

Provides a toggle to enable or disable it.