upgrade: cinder: run live migrations at correct rev
The Pike release of OpenStack Cinder requires the live migrations
in Ocata to be complete. A check in the database sync at version
97 enforces this.

This patch follows the conventions established in our Nova upgrade:

 1. db sync up to version 96
 2. Run the required live-migrations
 3. Complete the database sync

We are skipping the Ocata release and instead running the migrations
from a Pike codebase. However, Pike's sqlalchemy models include an
extra column to the groups database table that is therefore required
for the live migrations. Thus, we will copy the column creation from
db version 102 into a placeholder at 96. This will ensure the column
is created so that the Pike code is capable of running the live
migration.