rabbitmq: optimize for max connections
Allows marking a setting to optimize the underlying erlang and
kernel settings to deal with a higher number of connections,
especially with burst connections.
This will require more resources and may affect throughput.

Adds extra erlangs arguments:
 - +A: Sets the number of threads in async thread pool.
Valid range is 0-1024. Defaults to 128. Set to double of that.

 - +P: Sets the maximum number of simultaneously existing
processes for this system. Valid range for is [1024-134217727].
Set to half of maximum value.

Adds extra kernel params:
 - net.ipv4.ip_local_port_range: increase the range of ports available.
Default is 27018 to 64999. Set to 15000 to 65000

 - net.ipv4.tcp_fin_timeout: reduces the amount of time closed connections
will stay in the TIME_WAIT state. Default is 60. Set to 10

 - net.core.somaxconn: size of the listen queue at the same time.
Default is 128. Set to 4096.

 - net.ipv4.tcp_max_syn_backlog: maximum number of remembered connection
requests which did not receive an acknowledgment yet from connecting client.
Default is 128. Mainly for throughput but it may help with the burst.
Set to 4096.