WIP Allow both postgres+mariadb on the same node
This is to allow users to migrate from postgres to mariadb. It enables
multiple proposals for the database barclamp so that the deployer can
choose between migrating on the same node or over to a new node.

This assumes there are three scenarios:
 - 1 single instance of postgres is deployed (keep using in legacy)
 - 1 single instance of mariadb is deployed (migration finished)
 - 1 instance of each postgresql and mariadb is deployed. (In this case
postgres is still used by the control plane until the migration is
complete).

TODO:
 - Handle deploying to a new node
 - Handle deploying postgres+HA and mariadb+galera
 - Handle the 2 node drdb case
 - Ensure/enforce only the above 3 scenarios are possible. ie, disallow
a second postgres to be deployed.

This code is kinda hacky but only intended as a temporary measure for
the migration in cloud 7. This should not be in cloud 8.