database: Split database-server role into backend specific roles
This commit splits the "database-server" role into "postgresql-server"
and "mysql-server". This is the first change in a series to allow
deploying both databases from a single proposal to the same (or
different) set of nodes.

The current state still does only allow a single database to be
deployed, follow up commits with enhance the code to allow for parallel
deployments.

Schema migrations were added to reassign the roles correctly on existing
deployments.