neutron [cisco-aci]: Add cisco-aci role for neutron.
So far, the implementation assumed that only compute nodes
would be attached to the ACI leaf ports and the controller
nodes are independently attached to the network not connected
to the ACI fabric itself. The only form of communication from
the controller to the ACI is through the agents. However,
recent customer POCs have changed our assumption since these
setups expect all or most of the nodes attached the ACI leaf
as the ACI fabric forms the central network infrastructure
of the entire deployment. In such scenarios, we need a mechanism
where we can know which of the nodes are attached to the ACI
leaf and which nodes are not. The agents should be enabled
and run on each node attached to the ACI leaf ports.
This commit tries to achieve exactly this result, so that
future deployments can flexibly allow any node attached to
the fabric.
The commit also validates that at least one node is assigned
with neutron-sdn-cisco-aci-agents role if the mechanism drivers for
apic_ml2 or apic_gbp is chosen.

(cherry picked from commit a63b7346eea507024fa8308da5785ba72f6294cd)