mysql: create .my.cnf in root home directory for mysql cmdline
Searching for the mysql database root password can be a major
pain, so just store it in ~/.my.cnf so that mysql defaults
to it and everything just works when you're root on the controller.

As you can change the password when you're root, this shouldn't
be an additional security hassle.

(cherry picked from commit 95d2936fd5896a6942adcbb22cfa34c02f1fbec2)