Restore caching of db_settings
We need to cache settings for each combination of database barclamp
instances and sql engines. To do this, we first check (and cache) the
sql_engine selected by the role. The exception is when running the
cookbooks for the engines themselves.

(cherry picked from commit 824658f002ea5311fb6db885f9406df43ad6d864)