Set a proper ACL for shared certificates (bsc#1101154)
If the user is sharing a certificate between different OpenStack
services, this needs to be readed for different system users.  For
example, if the client store the certificate in /etc/apache2/ssl.key,
this directory and all the parents needs to be read for all the
OpenStack service users, and also the certificates contained there.

To address that we use the ACL from the operating system, via the
`setfacl` command.

(cherry picked from commit 75fe195ad7875fd650bb842d8127c930d87d5b6b)