Synchronize SSL in the cluster (bsc#1081518)
When Horizon in deployed via a autogenerated self certificate, the
current code will generate a different one of each node in the
cluster.

This patch will generate a single certificate in the founder, and
share it with the rest of the cluster via `rsync`.

(cherry picked from commit b033efa1d34a580f1326b7ab41a75195dfb965be)