keystone: Always try to rsync keys to new nodes
When adding new node to existing cluster, fernet keys need to be pushed from
the founder to keep all keystone nodes in sync. Before this change the keys
were pushed to all nodes when cluster was initially set up and then rotated
periodically.

This commit introduces internal node attribute which indicates that the keys
are already in place. Founder pushes the fernet keys to all nodes which don't
have this attribute set. The same code covers initial cluster setup as well
as adding new node to existing cluster.

In addition the lookup of cluster member nodes was changed. Helper function
from crowbar-pacemaker cookbook which was based on Chef search was replaced
with direct lookup of `elements` attribute (based on deployment part of
proposal). This way, new soon-to-be-cluster-member node will be included in
the list even before Chef search index is updated and/or chef-client is run
on that node. This guarantees that fernet keys are pushed to the new node as
soon as possible.

(based on commit 901d3bd4e00e438fa31d2e64dfa49c75f6cf8ddd)