nova: Increase HA resources creation timeout
When new node is added to existing cluster, all cloned resources need to be
extended to include the new node. This operation takes more time than initial
creation of resources as it is implemented as a series of "update" calls
rather than single transaction. Updated timeout was based on timing stats
collected from test runs.