horizon: Increase HA timeout
When new node is added to existing cluster, Horizon needs to be installed and
configured on that node. Because "old" nodes have most of required packages
and configurations already in place, this step caused the new node to fall
behind so much that the default sync window was breached.
Updated timeout was based on timing stats collected from test runs.