database: Prepare MariaDB for migration
Before data could be migrated from PostgreSQL databases and
users for all services need to be created in target MariaDB.
Also table structures need to be deployed (using OpenStack
manage commands).

New script (prepare-mariadb) was added to perform the whole
process. Role specific data was extracted from existing
pg2mariadb_preparation recipe so that it can be used in both
recipe and the prepare script.

The script looks up nodes which host services using the
database and prepares each of them.

First step is execution of pg2mariadb_preparation chef recipe
which creates all needed databases and users. It also writes
set of config override files and per-role scripts which are
later used to execute dbsync commands.
The recipe is executed by temporary modification of node-role
run_list and running chef-client on the target node.

Second step is execution of all dbsync scripts on the target
node.

The script should be executed on the admin server.
Stdout and stderr of all steps are written to log files
under /var/log/crowbar/db-prepare*.log