keystone: Reorganize HA/fernet code
Moved most of the fernet setup to ha.rb. Some redundancy was added but
it is now easier to work with the (more-complex) HA case.

The main goal was to reorder step so that fernet keys are configured and
distributed to cluster nodes before apache is (re)started. This way, there
is no time when apache would be configured to use fernet keys but no actual
keys would be present. The problem was detected in cluster extension scenario
where new node was added to an already existing cluster.