database: Execute db_sync commands via Chef
The execution of db_sync commands was moved from wrapper script to the
recipe. This way additional ssh calls could be removed and the wrapper
script is simplified.

In addition, the recipe is executed on first database node to create one
common index file which could be used as input for the migration tool.