rabbitmq: make notification queues optional
If notifications are enabled but ceilometer (or any other consumer like
monasca-ceilometer) isn't deployed, there isn't
any consumer for those messages and since they're configured to not
expire, they will just eternally pile up in the mnesia database. This
is a problem because in nontrivial environments this can be multiple
gigabytes a day of data that just keeps piling up slowing everything
down.

We could set a expiry policy on those notifications so that if there
isn't a consumer they disappear, but that still causes the overhead
and would be highly undesirable in case billing/telemetry is actually
intended to be used (those events must not ever get lost in such a
case even if the consumer is unavailable for some time).

For stable/ branch the backport was changed to not change the default
but just add an option to turn it off.

(cherry picked from commit 966090fac188bb9081e1a9cdc4a45ca30e2079ed)