kick off an experiment with CODEOWNERS
This is an experiment which could supersede the mentionbot
which was just removed.  It will cause the mentioned users
to be automatically notified when pull requests are proposed
which modify files matching the specified patterns, as described in

    https://help.github.com/articles/about-codeowners/

In our context, it does not mean anything about real ownership (yet),
only notification.

Users and file patterns have been picked without any serious thought,
and can of course easily be improved later.