rabbit: Change failover strategy to shuffle
The default, round-robin, leads to a predictable reconnect order
as we're listing the services always in the same order in the config
file. This causes queues to be be less balanced, e.g. the first
node is always servicing more load than the other nodes in the rabbitmq
cluster. with a shuffle strategy this should become more balanced over
time.