neutron: disable metadata proxy when metadata is forced
If 'force_metadata' is configured as 'True' then disable the metadata
proxies in the nodes do not have dhcp namespaces. They are not needed
in l3 agents.

And then 'enable_isolated_metadata' is neglected. So there is no need to
configure it, since 'force_metadata' is a superset and configured that
alone is sufficient to get the metadata from the dhcp server.

Also 'enable_metadata_network' may not be required if
'enable_isolated_metadata' is enabled, so it is disabled too.