rabbitmq: Remove address from node attributes
Storing the address in node attributes in rabbitmq cookbook and reading them in other cookbooks
(via CrowbarOpenStackHelper) can cause problems when recipe tries to use address from new node
before it sets it in the node.
Replacing this with direct address lookup based on Network barclamp attributes solves this problem.