Add the SES Barclamp
This padds the new SUSE Enterprise Storage (SES) barclamp.
This barclamp can accept the SES 5.5 openstack runner's
config.yml output.  The config.yml is added to a UI textbox
and then parsed into the required fields in the barclamp ui,
to support integration with an external SES/ceph cluster.

TODO:
  The only portion of the barclamp that currently works is the UI.
  Still have to add the required service node's system level packages.
  Still have to add the libvirt secret on all of the nova compute nodes.
  Still have to make the required changes to support the ceph backend
  settings in each of the service config files (cinder.conf, etc.)