rabbitmq: configure notification settings also in clustered mode
Accidentally when forward porting the patch the setting was
only set for the native clustered mode because the hash is
being assembled in two locations. Break out the common part
and merge it so that we follow DRY principles.

(based on commit a06bbb3663f311ac47b85e0fea64d6973c221d12)