ironic: Use ironic network for provisioning (bsc#1108398)
Switched to using "ironic" network IP in api_url which is passed to
ironic-python-agent running on the node during deployment and/or cleanup.
This should fix problems when IPA can't access ironic API because ironic
network is the only one attached to the node and there is no routed
access to public network via ironic network.