neutron: disable metering agent if no ceilometer
Currently we are enabling and running the metering agent at all
times in the network nodes for no reason.

Instead this patch makes it so its only deployed in the case of having
the ceilometer-agent role, which indicates that we are gathering
metrics and we want to run it. Otherwise its just wasted resources

(cherry picked from commit 17206688fe668ac5cba0517d4135155c9f5afd3b)