keystone: Remove UUID token format (SCRD-4086)
In Rocky the UUID token format was removed[1][2]. This patch removes
uuid as a valid token format, leaving only fernet. The dropdown menu is
also removed, however the option is still left in the schema to allow
for the possibility that new token formats may be added in the future.

[1] https://docs.openstack.org/releasenotes/keystone/rocky.html
[2] https://review.openstack.org/543060