monasca: install agent before run setup monitors in server
If one node has the role monasca-server it will use the agent
in spite this node doesn't have the monsaca-agent because the agent
installation is implicit in the monasca-server script.

But monasca-server skips the installation (And also the agent
installation)
if monasca-maser is not ready. It's only a warning. So Chef will
continue with the execution and when the others barclamps try to
prepare the monitoring fails.

With this change we ensure that the monasca-agent will be installed
in spite of the skip. And when the monasca-server will run its
installation
will reconfigure the agent.

Cherry-Picked from 7254cc1675d1041f039a31e9aa0aed17679993c8